requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

loguru==0.6.0
more-itertools==9.0.0
numpy==1.24.2
orjson==3.8.6
python-dotenv==0.21.1
sentence-transformers==2.2.2
